chore(opentelemetry source): config module and reduce duplication
