+++ src/sources/opentelemetry/tests.rs
@@ -6,7 +6,7 @@ use crate::{
     event::{
         into_event_stream, Event, EventStatus, LogEvent, Metric as MetricEvent, ObjectMap, Value,
     },
-    sources::opentelemetry::{GrpcConfig, HttpConfig, OpentelemetryConfig, LOGS, METRICS},
+    sources::opentelemetry::config::{GrpcConfig, HttpConfig, OpentelemetryConfig, LOGS, METRICS},
     test_util::{
         self,
         components::{assert_source_compliance, SOURCE_TAGS},
@@ -24,26 +24,74 @@ use std::time::{SystemTime, UNIX_EPOCH};
 use tonic::Request;
 use vector_lib::config::LogNamespace;
 use vector_lib::lookup::path;
-use vector_lib::opentelemetry::proto::collector::metrics::v1::metrics_service_client::MetricsServiceClient;
-use vector_lib::opentelemetry::proto::collector::metrics::v1::ExportMetricsServiceRequest;
-use vector_lib::opentelemetry::proto::common::v1::any_value::Value::StringValue;
-use vector_lib::opentelemetry::proto::metrics::v1::exponential_histogram_data_point::Buckets;
-use vector_lib::opentelemetry::proto::metrics::v1::metric::Data;
-use vector_lib::opentelemetry::proto::metrics::v1::summary_data_point::ValueAtQuantile;
-use vector_lib::opentelemetry::proto::metrics::v1::{
-    AggregationTemporality, ExponentialHistogram, ExponentialHistogramDataPoint, Gauge, Histogram,
-    HistogramDataPoint, Metric, NumberDataPoint, ResourceMetrics, ScopeMetrics, Sum, Summary,
-    SummaryDataPoint,
-};
-use vector_lib::opentelemetry::proto::resource::v1::Resource;
 use vector_lib::opentelemetry::proto::{
-    collector::logs::v1::{logs_service_client::LogsServiceClient, ExportLogsServiceRequest},
-    common::v1::{any_value, AnyValue, InstrumentationScope, KeyValue},
+    collector::{
+        logs::v1::{logs_service_client::LogsServiceClient, ExportLogsServiceRequest},
+        metrics::v1::{metrics_service_client::MetricsServiceClient, ExportMetricsServiceRequest},
+    },
+    common::v1::{any_value, any_value::Value::StringValue, AnyValue, InstrumentationScope, KeyValue},
     logs::v1::{LogRecord, ResourceLogs, ScopeLogs},
-    resource::v1::Resource as OtelResource,
+    metrics::v1::{
+        exponential_histogram_data_point::Buckets, metric::Data, summary_data_point::ValueAtQuantile, AggregationTemporality,
+        ExponentialHistogram, ExponentialHistogramDataPoint, Gauge, Histogram, HistogramDataPoint, Metric,
+        NumberDataPoint, ResourceMetrics, ScopeMetrics, Sum,
+        Summary, SummaryDataPoint,
+    },
+    resource::v1::{Resource, Resource as OtelResource},
 };
 use vrl::value;
 use warp::http::HeaderMap;
+
+fn create_test_logs_request() -> Request<ExportLogsServiceRequest> {
+    Request::new(ExportLogsServiceRequest {
+        resource_logs: vec![ResourceLogs {
+            resource: Some(OtelResource {
+                attributes: vec![KeyValue {
+                    key: "res_key".into(),
+                    value: Some(AnyValue {
+                        value: Some(StringValue("res_val".into())),
+                    }),
+                }],
+                dropped_attributes_count: 0,
+            }),
+            scope_logs: vec![ScopeLogs {
+                scope: Some(InstrumentationScope {
+                    name: "some.scope.name".into(),
+                    version: "1.2.3".into(),
+                    attributes: vec![KeyValue {
+                        key: "scope_attr".into(),
+                        value: Some(AnyValue {
+                            value: Some(StringValue("scope_val".into())),
+                        }),
+                    }],
+                    dropped_attributes_count: 7,
+                }),
+                log_records: vec![LogRecord {
+                    time_unix_nano: 1,
+                    observed_time_unix_nano: 2,
+                    severity_number: 9,
+                    severity_text: "info".into(),
+                    body: Some(AnyValue {
+                        value: Some(StringValue("log body".into())),
+                    }),
+                    attributes: vec![KeyValue {
+                        key: "attr_key".into(),
+                        value: Some(AnyValue {
+                            value: Some(StringValue("attr_val".into())),
+                        }),
+                    }],
+                    dropped_attributes_count: 3,
+                    flags: 4,
+                    // opentelemetry sdk will hex::decode the given trace_id and span_id
+                    trace_id: str_into_hex_bytes("4ac52aadf321c2e531db005df08792f5"),
+                    span_id: str_into_hex_bytes("0b9e4bda2a55530d"),
+                }],
+                schema_url: "v1".into(),
+            }],
+            schema_url: "v1".into(),
+        }],
+    })
+}
 
 #[test]
 fn generate_config() {
@@ -64,54 +112,7 @@ async fn receive_grpc_logs_vector_namespace() {
         let mut client = LogsServiceClient::connect(format!("http://{}", env.grpc_addr))
             .await
             .unwrap();
-        let req = Request::new(ExportLogsServiceRequest {
-            resource_logs: vec![ResourceLogs {
-                resource: Some(OtelResource {
-                    attributes: vec![KeyValue {
-                        key: "res_key".into(),
-                        value: Some(AnyValue {
-                            value: Some(StringValue("res_val".into())),
-                        }),
-                    }],
-                    dropped_attributes_count: 0,
-                }),
-                scope_logs: vec![ScopeLogs {
-                    scope: Some(InstrumentationScope {
-                        name: "some.scope.name".into(),
-                        version: "1.2.3".into(),
-                        attributes: vec![KeyValue {
-                            key: "scope_attr".into(),
-                            value: Some(AnyValue {
-                                value: Some(StringValue("scope_val".into())),
-                            }),
-                        }],
-                        dropped_attributes_count: 7,
-                    }),
-                    log_records: vec![LogRecord {
-                        time_unix_nano: 1,
-                        observed_time_unix_nano: 2,
-                        severity_number: 9,
-                        severity_text: "info".into(),
-                        body: Some(AnyValue {
-                            value: Some(StringValue("log body".into())),
-                        }),
-                        attributes: vec![KeyValue {
-                            key: "attr_key".into(),
-                            value: Some(AnyValue {
-                                value: Some(StringValue("attr_val".into())),
-                            }),
-                        }],
-                        dropped_attributes_count: 3,
-                        flags: 4,
-                        // opentelemetry sdk will hex::decode the given trace_id and span_id
-                        trace_id: str_into_hex_bytes("4ac52aadf321c2e531db005df08792f5"),
-                        span_id: str_into_hex_bytes("0b9e4bda2a55530d"),
-                    }],
-                    schema_url: "v1".into(),
-                }],
-                schema_url: "v1".into(),
-            }],
-        });
+        let req = create_test_logs_request();
         _ = client.export(req).await;
         let mut output = test_util::collect_ready(env.output).await;
         // we just send one, so only one output
@@ -209,54 +210,7 @@ async fn receive_grpc_logs_legacy_namespace() {
         let mut client = LogsServiceClient::connect(format!("http://{}", env.grpc_addr))
             .await
             .unwrap();
-        let req = Request::new(ExportLogsServiceRequest {
-            resource_logs: vec![ResourceLogs {
-                resource: Some(OtelResource {
-                    attributes: vec![KeyValue {
-                        key: "res_key".into(),
-                        value: Some(AnyValue {
-                            value: Some(StringValue("res_val".into())),
-                        }),
-                    }],
-                    dropped_attributes_count: 0,
-                }),
-                scope_logs: vec![ScopeLogs {
-                    scope: Some(InstrumentationScope {
-                        name: "some.scope.name".into(),
-                        version: "1.2.3".into(),
-                        attributes: vec![KeyValue {
-                            key: "scope_attr".into(),
-                            value: Some(AnyValue {
-                                value: Some(StringValue("scope_val".into())),
-                            }),
-                        }],
-                        dropped_attributes_count: 7,
-                    }),
-                    log_records: vec![LogRecord {
-                        time_unix_nano: 1,
-                        observed_time_unix_nano: 2,
-                        severity_number: 9,
-                        severity_text: "info".into(),
-                        body: Some(AnyValue {
-                            value: Some(StringValue("log body".into())),
-                        }),
-                        attributes: vec![KeyValue {
-                            key: "attr_key".into(),
-                            value: Some(AnyValue {
-                                value: Some(StringValue("attr_val".into())),
-                            }),
-                        }],
-                        dropped_attributes_count: 3,
-                        flags: 4,
-                        // opentelemetry sdk will hex::decode the given trace_id and span_id
-                        trace_id: str_into_hex_bytes("4ac52aadf321c2e531db005df08792f5"),
-                        span_id: str_into_hex_bytes("0b9e4bda2a55530d"),
-                    }],
-                    schema_url: "v1".into(),
-                }],
-                schema_url: "v1".into(),
-            }],
-        });
+        let req = create_test_logs_request();
         _ = client.export(req).await;
         let mut output = test_util::collect_ready(env.output).await;
         // we just send one, so only one output
